Extract gestaltd daemon entrypoint
Introduce the internal daemon boundary:

    type Options struct { Version string; Args []string }

    func Run(opts Options) error

Example usage from cmd/gestaltd:

    daemon.Run(daemon.Options{Version: version, Args: os.Args[1:]})

Move serve/lock/sync/init/validate/provider command implementation and existing command integration tests under internal/daemon while cmd/gestaltd remains the binary wrapper. Keep daemon internals private to the gestaltd module via Go's internal package boundary.
