+++ gestaltd/internal/daemon/init.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 import (
 	"context"
