+++ gestaltd/internal/daemon/main_test.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 import (
 	"os/exec"
