+++ gestaltd/internal/daemon/provider.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 import (
 	"bytes"
