+++ gestaltd/internal/daemon/provider_attach.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 import (
 	"context"
