+++ gestaltd/internal/daemon/provider_test.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 import (
 	"context"
